requirements.txt
fastapi
uvicorn
pandas
pandas_ta
google-generativeai
pydantic